# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "HubGit Provider API",
    "version": "0.1.0",
    "description": "Provider-neutral contract implemented by HubGit mock and future Git host adapters."
  },
  "servers": [{ "url": "/", "description": "Current HubGit instance" }],
  "security": [{ "sessionCookie": [] }],
...
  "paths": {
...
    "/api/v1/viewer": {
      "get": { "tags": ["identity"], "operationId": "getViewer" , "summary": "get viewer", "responses": { "200": { "$ref": "#/components/responses/Viewer" }, "401": { "$ref": "#/components/responses/Problem" } } }
    },
...
    "/api/v1/auth/session": {
      "get": { "tags": ["auth"], "operationId": "getSession" , "summary": "get session", "security": [], "responses": { "200": { "$ref": "#/components/responses/Session" } } },
      "delete": { "tags": ["auth"], "operationId": "logout" , "summary": "logout", "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }], "responses": { "204": { "description": "Logged out" }, "403": { "$ref": "#/components/responses/Problem" } } }
    },
    "/api/v1/auth/login": {
      "post": { "tags": ["auth"], "operationId": "login" , "summary": "login", "security": [], "requestBody": { "$ref": "#/components/requestBodies/Login" }, "responses": { "200": { "$ref": "#/components/responses/Session" }, "401": { "$ref": "#/components/responses/Problem" }, "422": { "$ref": "#/components/responses/Problem" } } }
    },
...
    "/api/v1/repositories": {
      "get": { "tags": ["repositories"], "operationId": "listRepositories" , "summary": "list repositories", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Query" }], "responses": { "200": { "$ref": "#/components/responses/RepositoryPage" }, "401": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" }, "503": { "$ref": "#/components/responses/Problem" } } },
      "post": { "tags": ["repositories"], "operationId": "createRepository" , "summary": "create repository", "parameters": [{ "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IdempotencyKey" }], "requestBody": { "$ref": "#/components/requestBodies/RepositoryCreate" }, "responses": { "201": { "$ref": "#/components/responses/Repository" }, "409": { "$ref": "#/components/responses/Problem" }, "422": { "$ref": "#/components/responses/Problem" } } }
    },
    "/api/v1/repositories/{owner}/{repo}": {
      "get": { "tags": ["repositories"], "operationId": "getRepository" , "summary": "get repository", "security": [], "parameters": [{ "$ref": "#/components/parameters/Owner" }, { "$ref": "#/components/parameters/Repo" }], "responses": { "200": { "$ref": "#/components/responses/Repository" }, "401": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" }, "503": { "$ref": "#/components/responses/Problem" } } },
      "patch": { "tags": ["settings"], "operationId": "updateRepository" , "summary": "update repository", "parameters": [{ "$ref": "#/components/parameters/Owner" }, { "$ref": "#/components/parameters/Repo" }, { "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IfMatch" }], "requestBody": { "$ref": "#/components/requestBodies/RepositoryUpdate" }, "responses": { "200": { "$ref": "#/components/responses/Repository" }, "412": { "$ref": "#/components/responses/Problem" } } },
      "delete": { "tags": ["settings"], "operationId": "deleteRepository" , "summary": "delete repository", "parameters": [{ "$ref": "#/components/parameters/Owner" }, { "$ref": "#/components/parameters/Repo" }, { "$ref": "#/components/parameters/CsrfToken" }, { "$ref": "#/components/parameters/IfMatch" }], "responses": { "204": { "description": "Repository deleted" }, "412": { "$ref": "#/components/responses/Problem" } } }
    },
...
    "/api/v1/repositories/{owner}/{repo}/tree/{ref}": {
      "get": { "tags": ["git"], "operationId": "getTree" , "summary": "get tree", "security": [], "parameters": [{ "$ref": "#/components/parameters/Owner" }, { "$ref": "#/components/parameters/Repo" }, { "$ref": "#/components/parameters/Ref" }, { "$ref": "#/components/parameters/Path" }], "responses": { "200": { "description": "Git tree", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GitTree" } } } }, "401": { "$ref": "#/components/responses/Problem" }, "404": { "$ref": "#/components/responses/Problem" }, "429": { "$ref": "#/components/responses/Problem" }, "503": { "$ref": "#/components/responses/Problem" } } }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "hubgit_session" }
    },
    "headers": {
      "ETag": { "description": "Opaque resource version. Treat as an opaque quoted value and send it unchanged in If-Match.", "schema": { "type": "string", "minLength": 1 } },
      "NoStore": { "description": "Sensitive response must not be stored by shared caches", "schema": { "type": "string", "const": "no-store" } },
...
    },
    "parameters": {
      "Owner": { "name": "owner", "in": "path", "required": true, "schema": { "type": "string" } },
      "Repo": { "name": "repo", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "Ref": { "name": "ref", "in": "path", "required": true, "schema": { "type": "string" } },
...
      "Path": { "name": "path", "in": "query", "schema": { "type": "string", "default": "" } },
...
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string" } },
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 30 } },
      "Query": { "name": "q", "in": "query", "schema": { "type": "string" } },
...
      "CsrfToken": { "name": "X-CSRF-Token", "in": "header", "required": true, "schema": { "type": "string" } },
      "IfMatch": { "name": "If-Match", "in": "header", "required": true, "schema": { "type": "string" } },
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "type": "string", "minLength": 16, "maxLength": 128 } }
    },
    "requestBodies": {
      "Login": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } } },
...
      "RepositoryCreate": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepositoryCreate" } } } },
      "RepositoryUpdate": { "required": true, "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/RepositoryUpdate" } } } },
...
    },
    "responses": {
      "Problem": { "description": "Problem details", "headers": { "Cache-Control": { "$ref": "#/components/headers/NoStore" } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/ProblemDetails" } } } },
...
      "Viewer": { "description": "Current viewer", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Viewer" } } } },
      "Session": { "description": "Authentication session", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Session" } } } },
...
      "Repository": { "description": "Repository", "headers": { "ETag": { "$ref": "#/components/headers/ETag" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Repository" } } } },
      "RepositoryPage": { "description": "Repository page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/RepositoryPage" } } } },
...
    },
    "schemas": {
      "ProblemDetails": {
        "type": "object", "required": ["type", "title", "status", "code"],
        "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer" }, "detail": { "type": "string" }, "instance": { "type": "string", "format": "uri-reference" }, "code": { "type": "string", "examples": ["repository.not_found"] }, "fieldErrors": { "type": "array", "items": { "$ref": "#/components/schemas/FieldError" } }, "traceId": { "type": "string" } }
      },
      "FieldError": { "type": "object", "required": ["field", "code", "message"], "properties": { "field": { "type": "string" }, "code": { "type": "string" }, "message": { "type": "string" } } },
      "PageInfo": { "type": "object", "required": ["hasNextPage", "hasPreviousPage"], "properties": { "startCursor": { "type": ["string", "null"] }, "endCursor": { "type": ["string", "null"] }, "hasNextPage": { "type": "boolean" }, "hasPreviousPage": { "type": "boolean" } } },
      "ResourcePermissions": { "type": "object", "required": ["read", "write", "admin"], "properties": { "read": { "type": "boolean" }, "triage": { "type": "boolean", "default": false }, "write": { "type": "boolean" }, "maintain": { "type": "boolean", "default": false }, "admin": { "type": "boolean" } }, "additionalProperties": false },
...
      "Freshness": { "type": "object", "description": "Canonical synchronization and authorization metadata for data that may be served from an adapter cache.", "required": ["state", "lastSyncedAt", "lastAuthorizedAt", "provider"], "properties": { "state": { "type": "string", "enum": ["live", "refreshing", "stale", "offline"] }, "lastSyncedAt": { "type": ["string", "null"], "format": "date-time" }, "lastAuthorizedAt": { "type": ["string", "null"], "format": "date-time" }, "provider": { "type": "string" } }, "additionalProperties": false },
...
      "LoginRequest": { "type": "object", "required": ["login", "password"], "properties": { "login": { "type": "string" }, "password": { "type": "string" }, "otp": { "type": "string", "pattern": "^[0-9]{6}$" }, "remember": { "type": "boolean", "default": false } } },
...
      "Session": { "type": "object", "required": ["authenticated", "csrfToken", "expiresAt", "viewer"], "properties": { "authenticated": { "type": "boolean" }, "csrfToken": { "type": "string" }, "expiresAt": { "type": ["string", "null"], "format": "date-time" }, "viewer": { "oneOf": [{ "$ref": "#/components/schemas/Viewer" }, { "type": "null" }] } } },
      "Viewer": { "allOf": [{ "$ref": "#/components/schemas/User" }, { "type": "object", "required": ["roles"], "properties": { "roles": { "type": "array", "items": { "type": "string", "enum": ["member", "maintainer", "administrator"] }, "uniqueItems": true }, "email": { "type": "string", "format": "email" }, "unreadNotificationCount": { "type": "integer", "minimum": 0 } } }] },
      "User": { "type": "object", "required": ["id", "kind", "username", "displayName", "avatarUrl", "profileUrl"], "properties": { "id": { "type": "string" }, "kind": { "const": "user" }, "username": { "type": "string" }, "displayName": { "type": "string" }, "avatarUrl": { "type": "string", "format": "uri-reference" }, "profileUrl": { "type": "string", "format": "uri-reference" }, "bio": { "type": ["string", "null"] }, "location": { "type": ["string", "null"] }, "followerCount": { "type": "integer", "minimum": 0 }, "followingCount": { "type": "integer", "minimum": 0 }, "createdAt": { "type": "string", "format": "date-time" }, "freshness": { "$ref": "#/components/schemas/Freshness" } } },
...
      "Repository": {
        "type": "object", "required": ["id", "kind", "owner", "name", "fullName", "visibility", "defaultBranch", "empty", "archived", "permissions", "counts", "cloneUrls", "createdAt", "updatedAt"],
        "properties": { "id": { "type": "string" }, "kind": { "const": "repository" }, "owner": { "$ref": "#/components/schemas/RepositoryOwner" }, "name": { "type": "string" }, "fullName": { "type": "string" }, "description": { "type": ["string", "null"] }, "visibility": { "type": "string", "enum": ["public", "internal", "private"] }, "defaultBranch": { "type": "string" }, "empty": { "type": "boolean" }, "archived": { "type": "boolean" }, "fork": { "type": "boolean", "default": false }, "forkedFrom": { "type": ["string", "null"] }, "language": { "type": ["string", "null"] }, "license": { "type": ["string", "null"] }, "topics": { "type": "array", "items": { "type": "string" } }, "permissions": { "$ref": "#/components/schemas/ResourcePermissions" }, "counts": { "$ref": "#/components/schemas/RepositoryCounts" }, "cloneUrls": { "$ref": "#/components/schemas/CloneUrls" }, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }, "pushedAt": { "type": ["string", "null"], "format": "date-time" }, "freshness": { "$ref": "#/components/schemas/Freshness" } }
      },
      "RepositoryOwner": { "type": "object", "required": ["id", "kind", "login", "avatarUrl"], "properties": { "id": { "type": "string" }, "kind": { "type": "string", "enum": ["user", "organization"] }, "login": { "type": "string" }, "avatarUrl": { "type": "string", "format": "uri-reference" } } },
      "RepositoryCounts": { "type": "object", "required": ["stars", "forks", "watchers", "issues", "pullRequests"], "properties": { "stars": { "type": "integer" }, "forks": { "type": "integer" }, "watchers": { "type": "integer" }, "issues": { "type": "integer" }, "pullRequests": { "type": "integer" } } },
      "CloneUrls": { "type": "object", "required": ["http", "ssh"], "properties": { "http": { "type": "string", "format": "uri-reference" }, "ssh": { "type": "string" } } },
      "RepositoryCreate": { "type": "object", "required": ["owner", "name", "visibility"], "properties": { "owner": { "type": "string" }, "name": { "type": "string", "pattern": "^[A-Za-z0-9._-]+$" }, "description": { "type": ["string", "null"], "maxLength": 350 }, "visibility": { "type": "string", "enum": ["public", "internal", "private"] }, "initialize": { "type": "boolean", "default": false } } },
      "RepositoryUpdate": { "type": "object", "properties": { "name": { "type": "string" }, "description": { "type": ["string", "null"] }, "visibility": { "type": "string", "enum": ["public", "internal", "private"] }, "defaultBranch": { "type": "string" }, "archived": { "type": "boolean" }, "features": { "$ref": "#/components/schemas/RepositoryFeatures" } } },
      "RepositoryFeatures": { "type": "object", "properties": { "issues": { "type": "boolean" }, "wiki": { "type": "boolean" }, "discussions": { "type": "boolean" }, "projects": { "type": "boolean" } } },
      "RepositoryPage": { "type": "object", "required": ["items", "pageInfo"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Repository" } }, "pageInfo": { "$ref": "#/components/schemas/PageInfo" }, "totalCount": { "type": "integer" } } },
...
      "GitTree": { "type": "object", "required": ["sha", "ref", "path", "entries", "commit"], "properties": { "sha": { "type": "string" }, "ref": { "type": "string" }, "path": { "type": "string" }, "entries": { "type": "array", "items": { "$ref": "#/components/schemas/TreeEntry" } }, "commit": { "$ref": "#/components/schemas/CommitSummary" } } },
      "TreeEntry": { "type": "object", "required": ["name", "path", "kind", "sha"], "properties": { "name": { "type": "string" }, "path": { "type": "string" }, "kind": { "type": "string", "enum": ["file", "directory", "symlink", "submodule"] }, "sha": { "type": "string" }, "size": { "type": ["integer", "null"] }, "lastCommit": { "$ref": "#/components/schemas/CommitSummary" } } },
...
      "CommitSummary": { "type": "object", "required": ["sha", "shortSha", "message", "author", "authoredAt"], "properties": { "sha": { "type": "string" }, "shortSha": { "type": "string" }, "message": { "type": "string" }, "author": { "$ref": "#/components/schemas/GitActor" }, "authoredAt": { "type": "string", "format": "date-time" }, "verification": { "$ref": "#/components/schemas/Verification" } } },
      "GitActor": { "type": "object", "required": ["name", "email"], "properties": { "name": { "type": "string" }, "email": { "type": "string", "format": "email" }, "user": { "oneOf": [{ "$ref": "#/components/schemas/User" }, { "type": "null" }] } } },
      "Verification": { "type": "object", "required": ["verified"], "properties": { "verified": { "type": "boolean" }, "reason": { "type": ["string", "null"] }, "signatureType": { "type": ["string", "null"], "enum": ["gpg", "ssh", "smime", null] } } },
...
    }
  }
}